// apps/pharmacy/src/types/staff/doctor.ts
export interface CreateOrUpdateDoctor {
  id?: number;
  name: string;
  surname: string;
  employeeId: string;
  dob: string;
  title?: string | null;
  phone?: string | null;
  email: string;
  notes?: string | null;
  siteId?: number | null;
  designationId?: number | null;
  departmentId?: number | null;
  locationId?: number | null;
}